LAC-159: hbl permissions applied

+++ app/Http/Controllers/HBLController.php
@@ -13,12 +13,15 @@ use App\Interfaces\PriceRepositoryInterface;
 use App\Interfaces\UserRepositoryInterface;
 use App\Models\HBL;
 use App\Models\HBLDocument;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class HBLController extends Controller
 {
+    use AuthorizesRequests;
+
     public function __construct(
         private readonly HBLRepositoryInterface $HBLRepository,
         private readonly UserRepositoryInterface $userRepository,
@@ -31,6 +34,8 @@ class HBLController extends Controller
      */
     public function index()
     {
+        $this->authorize('hbls.index');
+
         return Inertia::render('HBL/HBLList', [
             'users' => $this->userRepository->getUsers(),
             'hbls' => $this->HBLRepository->getHBLsWithPackages(),
@@ -56,6 +61,8 @@ class HBLController extends Controller
      */
     public function create()
     {
+        $this->authorize('hbls.create');
+
         return Inertia::render('HBL/CreateHBL', [
             'cargoTypes' => CargoType::cases(),
             'hblTypes' => HBLType::cases(),
@@ -86,6 +93,8 @@ class HBLController extends Controller
      */
     public function edit(HBL $hbl)
     {
+        $this->authorize('hbls.edit');
+
         return Inertia::render('HBL/EditHBL', [
             'hbl' => $hbl->load('packages'),
             'cargoTypes' => CargoType::cases(),
@@ -108,21 +117,29 @@ class HBLController extends Controller
      */
     public function destroy(HBL $hbl)
     {
+        $this->authorize('hbls.delete');
+
         $this->HBLRepository->deleteHBL($hbl);
     }
 
     public function toggleHold(HBL $hbl)
     {
+        $this->authorize('hbls.hold and release');
+
         $this->HBLRepository->toggleHold($hbl);
     }
 
     public function downloadHBLPDF(HBL $hbl)
     {
+        $this->authorize('hbls.download pdf');
+
         return $this->HBLRepository->downloadHBLPDF($hbl);
     }
 
     public function cancelledHBLs()
     {
+        $this->authorize('hbls.show cancelled hbls');
+
         return Inertia::render('HBL/CancelledHBLList', [
             'users' => $this->userRepository->getUsers(),
             'hbls' => $this->HBLRepository->getHBLsWithPackages(),
@@ -145,16 +162,22 @@ class HBLController extends Controller
 
     public function restore($id)
     {
+        $this->authorize('hbls.restore');
+
         return $this->HBLRepository->restore($id);
     }
 
     public function downloadHBLInvoicePDF(HBL $hbl)
     {
+        $this->authorize('hbls.download invoice');
+
         return $this->HBLRepository->downloadHBLInvoicePDF($hbl);
     }
 
     public function downloadHBLBarcodePDF(HBL $hbl)
     {
+        $this->authorize('hbls.download barcode');
+
         return $this->HBLRepository->downloadHBLBarcodePDF($hbl);
     }
 
@@ -165,6 +188,8 @@ class HBLController extends Controller
 
     public function uploadDocument(Request $request)
     {
+        $this->authorize('hbls.upload documents');
+
         return $this->HBLRepository->uploadDocument($request->all());
     }
 
@@ -175,6 +200,8 @@ class HBLController extends Controller
 
     public function destroyHBLDocument(HBLDocument $hblDocument)
     {
+        $this->authorize('hbls.delete documents');
+
         $this->HBLRepository->deleteDocument($hblDocument);
     }
